requirements.txt
ccxt
ta-lib
python-telegram-bot
numpy
python-dotenv